requirements.txt
fastapi
uvicorn
pydantic-ai
python-multipart
PyPDF2
httpx
python-dotenv
